pyproject.toml
[project]
name = "data-loaders"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "pandas>=2.3.1",
    "psycopg2>=2.9.10",
    "python-dotenv>=1.1.1",
    "requests>=2.32.4",
    "ruff>=0.12.8",
]
